Make the wheel install target force an upgrade
Without better handling of version numbers, the
install target will not upgrade an existing wheel
correctly. This commit changes the install target
to explicitly remove the current versions of the
wheel. Not ideal for production (but for production
you're just going to install the wheels using pip),
but very helpful for testing.

Signed-off-by: Mic Bowman <mic.bowman@intel.com>