Allow for local override of contract families to be built
By default, all contract families are built. Tests will be run from
all families.

This update supports a local cmake file (Local.cmake) that will allow
for local overrides of variables. In particular, this is useful to set
the new CMake variable CONTRACT_FAMILIES to just the subset that you
want to build.

Signed-off-by: Mic Bowman <mic.bowman@intel.com>