add support for docker-based testing and execution
Building on the structure used for PDO, this commit adds a recipe for
building a contracts container that can be used for testing of the
contracts and the jupyter interface to the contracts.

The structure is similar to the PDO docker support: keys and
configuration variables are shared through the xfer directory. Tools
run in the container build the contracts, run the tests, and also run
a jupyter server.

The behavior can be customized through a local makefile that allows
for override of variables. A local environment file can be used to
customize the behavior docker compose.

Signed-off-by: Mic Bowman <mic.bowman@intel.com>