Variety of updates for the Jupyter
Several updates...

First, modified the CMakefiles to copy jupyter notebooks from the
different contract families into the ${PDO_JUPYTER_ROOT} directory as
part of the installation process.

Since the notebooks are now installed during build, we no longer need
to copy the them as part of the setup for the jupyter docker tests or
service. Removed the code to copy the notebooks.

Separated the guardian service out into a container of its own when
running the Jupyter tests. This is ultimately how the guardian should
be run. Added a script to start the guardian in the docker tools.

Signed-off-by: Mic Bowman <mic.bowman@intel.com>