Fix a problem with missing keys for guardian service
Make sure that necessary keys are available for the guardian
service and storage service.

Add an interface for the guardian service that tests for
the existance of a service. This was necessary (rather than
the straight URL tests we do for [eps]services) because the
configuration file cannot easily be parsed. We should probably
do this for all of the PDO services.

Signed-off-by: Mic Bowman <mic.bowman@intel.com>