Convert Jupyter notebooks to jupytext percent format (or markdown)
Jupyter notebooks are notoriously difficult to manage through version
control. The percent format moves the notebook to a more traditional
text format (e.g. python with markdown embedded or markdown with python
embedded). The Jupytext tool can convert from the percent formats to
notebooks and vice versa.

This commit replaces all notebook files with the percent version. And provides
cmake support for building and installing the notebook versions. That is, what
we run through jupyter will be notebooks generated from the percent version.

And... added the installation of jupytext and ipywidgets to the docker image.

Signed-off-by: Mic Bowman <mic.bowman@intel.com>