Update the getting started notebook
Replace the "documentation only" form of the getting started notebook
with one that provides the functionality for actually checking and
fixing the PDO configuration. The new form walks the user through key
creation, service endpoint import, and service group creation. This
should ensure that there is a fully configured environment for testing
services.

Signed-off-by: Mic Bowman <mic.bowman@intel.com>