Switch notebooks from service "host" to service "group"
Update the notebooks to take advantage of the transition to service
groups. This should enable significantly more flexibility in selecting
services from multiple sites.

Changed some of the way the "root" of a context is handled in order to
keep group specification easier. If all of the services groups
(eservice, pservice, and sservice) have the same name then a single
service group setting is sufficient to pick all. Otherwise, each group
can be overridden in the custom bindings.

Fixed line length in several of the notebooks.

Signed-off-by: Mic Bowman <mic.bowman@intel.com>