Enable token storage in asset wallet notebooks
Expose and use the necessary interfaces for tokens to be included in
wallet notebooks. The big benefit of this is that all "ownership"
functions can now be handled through the same interface whether they
are fungible assets (like marbles) or non-fungible assets (like
tokens).

Needed to support the "get_balance" method in the token contract (we
already support many of the other issuer methods through
tokens). Added better support for the balance and transfer in the
plugins.

And... most notably, added an "issuer" reference in the context
file. Context references turn out to be rather useful. This creates a
reference to the specific token object in the collections file. So now
there is a canonical reference point for accessing issuer functions.

Signed-off-by: Mic Bowman <mic.bowman@intel.com>