Address some memory issues in the digital asset contracts
Add some checks for memory allocation problems. This happens because
WAMR as it exists cannot generate/handle exceptions so C++ object allocation
can be problematic.

Change from bulk transfer of the object in the JSON parameters to a key/value
store for transfering data. This makes it possible to transfer binary image
data rather than go through the JSON encoding/decoding layers that use a lot
of memory.

Signed-off-by: Mic Bowman <mic.bowman@intel.com>