Code commit for example contract family
Initial commit for the example contract family. The example contract
family is primarily used as a means of documenting the process of
writing, building, and extending a contract family. In addition, the
example contract family provides a template for building a new
contract family. Copying the example contract family to a new
directory and editing family.cmake should be sufficient to create a
new family.

Signed-off-by: Mic Bowman <mic.bowman@intel.com>