Add Jupyter support for the example contract
Adds the basic factory/template model for the example contract. This
should allow for, more or less, a cut and paste for creating a new
contract that preserved the counter. That is, the factory & template
are for the specific contract type and would have to be replaced
completely for a new contract.

Signed-off-by: Mic Bowman <mic.bowman@intel.com>