Restructure the identity contract to match the structure used in the example contract
Move all source into a new "src" directory.

Move openssl into a packages directory in src.

Move all header files into src/identity (this should make it easier
for other contract families to pick up the methods and modules defined
in the identity cf.

Update the makefiles. The family.cmake file contains the name of the
contract family and, in theory, should now be the only file that
actually needs to be changed to rename/relocate the identity cf.

Signed-off-by: Mic Bowman <mic.bowman@intel.com>