feat(guardian, exchange): enhance request handling and add unique requests
- Refined error responses in guardian service and process capability apps
- Added support for request uniqueness checks in guardian operations
- Introduced `request_identifier` field to ensure unique operation requests
- Implemented a method to retrieve token identity in exchange token object
- Extended `create_operation_package` to include `request_identifier` handling

This is a very simple implementation of at-most-once semantics for
guardian capability processing. To use the feature, add a property to
the guardian capability handler class: "unique_request = True".  This
implementation uses an in memory set on a per-token basis. The
implemmentation uses the token identity property that is part of the
token-guardian protocol. Added a method to expose the token identity
value from the base token_object method class.

Signed-off-by: Mic Bowman <mic.bowman@intel.com>